Prevent infinite fetches, and cleanup of code
